Feat: Añadir la funcionalidad de buscar por min y max

+++ searchService/DTO´s/GradeDTO.cs
@@ -8,7 +8,7 @@ namespace searchService.DTOs
     public class GradeDTO
     {
         public string UUID { get; set; } = null!;
-        public int grade { get; set; } = 0!;
+        public float grade { get; set; } = 0!;
         public string gradeName { get; set; } = null!;
         public string subjectName { get; set; } = null!;
         public string comment { get; set; } = null!;
